:recycle: UserDao interface분리
