+++ src/main/java/com/msg/after_school/global/user/dao/UserDao.java
@@ -1,24 +1,8 @@
 package com.msg.after_school.global.user.dao;
 
 import com.msg.after_school.domain.user.data.entity.User;
-import com.msg.after_school.global.user.exception.UserNotFoundException;
-import com.msg.after_school.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
 
-@Component
-@RequiredArgsConstructor
-//TODO 인터페이스로 바꾸기 (안에 있는 구현체는 UserDaoImpl)
-public class UserDao {
-    private final UserRepository userRepository;
-
-    public User getCurrentUser() {
-        String email = SecurityContextHolder.getContext().getAuthentication().getName();
-        return getUserByEmail(email);
-    }
-
-    public User getUserByEmail(String email) {
-        return userRepository.findUserByEmail(email).orElseThrow(() -> new UserNotFoundException());
-    }
+public interface UserDao {
+    User getCurrentUser();
+    User getUserByEmail(String email);
 }
